requirements.txt
asgiref==3.8.1
Django==4.2.14
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
flake8-docstrings==1.7.0
inflection==0.5.1
mccabe==0.7.0
packaging==24.1
phonenumberslite==8.13.40
pillow==10.4.0
pycodestyle==2.12.0
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
snowballstemmer==2.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
